fix: reduced-precision islands for the belief kernel and transport (F0-F2)
F0 (both audits) -- check_audit_fixes.py referenced tests/test_belief_cache.py,
deleted with the active-inference/EFE removal in 53e72f1. Every node ID goes to
one pytest invocation, so the dangling path aborted collection for the whole
driver: exit 4, zero of the other 25 pinned regression tests run, banner reading
"FAILURES". Now collects 26 tests, exit 0.

F1 (both audits, independently) -- belief_gradients and mm_exact_update had no
float32 island under AMP while their own reference oracle re-enters fp32
(oracle.py:140). Under amp_dtype='bf16' the two sides of a seam the golden tests
pin at fp32 tolerance disagreed by ~4e-3 relative, the raw-energy saturation mask
that gates the pair derivative flipped on ~22% of pairs, the structural self-pair
energy stopped being exactly zero, and reuse_pairwise_kl_stats silently
self-disabled because its eligibility test requires fp32. Measured after: kernel
vs oracle |grad_mu| agreement 1.788e-07 under bf16 autocast, identical to the
pure-fp32 path.

F2 -- a represented group element is never STORED below float32, and its algebra
element is never BUILT below float32. Rounding the exponential and its float64
inverse back to the autocast dtype made U U^-1 deviate from the identity like
eps * cond(U), reaching O(1) near cond ~ 1/eps_bf16 ~ 128, and the vertex
factor's cocycle property IS Regime-I flatness. The edge exponential already
pinned fp32; this is the matching floor for the vertex. Measured at the phi_scale
init, bf16 autocast, before -> after: cocycle residual 1.32e-02 -> 2.30e-07,
||U U^-1 - I|| 7.56e-03 -> 3.96e-08, ||Omega_ii - I|| 3.91e-03 -> 1.19e-07. This
also closes the certificate blind spot for free: diagnostics rebuild outside
autocast, so their fp32 operator now IS the one the forward used.

Also F2: group_element_inverse's residual_tol was validated and then discarded,
leaving the one named guard against reduced-precision group elements inert. It
now bounds ||U U^-1 - I||_inf and raises, opt-in so no host sync lands on the hot
path. And the exponential's Frobenius clamp substitutes exp(M*scale) for exp(M)
-- a different operator, ~98% relative error at ||M||_F = 30 -- which was
announced only under the opt-in clamp_monitor; the first occurrence is now always
reported. transport_chart_max_norm's default is deliberately unchanged: every
value the pullback_group validator admits (5.0 < bound < 20.0) would hard-abort
runs whose observed ||M||_F already reaches 18.

Test contract updated, not loosened: the dense/compact dtype-parity pin now
expects float32 on both routes. Parity is preserved; it holds at fp32 instead of
bf16.

Full CPU lane read from JUnit XML: tests=4217 failures=9 errors=0 skipped=28.
The 9 are the identical set failing before these changes, all attributable to
uncommitted working-tree config edits (rope_on_value, e_step_update, a truncated
validator warning). Zero regressions.

+++ tests/test_p1_compact_phi_block_transport_20260711.py
@@ -570,7 +570,12 @@ def test_compact_phi_block_transport_matches_legacy_autocast_dtype_and_vjp() -> 
         outputs.append((mean, covariance))
         gradients.append(torch.autograd.grad(loss, phi)[0])
 
-    assert factor_dtypes == [torch.bfloat16, torch.bfloat16]
+    # Audit 2026-07-25 F2: the represented group element is never STORED below float32, on EITHER
+    # route. Rounding the exponential (and its float64 inverse) back to the autocast dtype made
+    # U U^-1 deviate from the identity like eps * cond(U) -- reaching O(1) near cond ~ 1/eps_bf16 --
+    # and the vertex factor's cocycle property IS Regime-I flatness. This test's contract is dense/
+    # compact PARITY, which is preserved; the parity now holds at float32 instead of at bfloat16.
+    assert factor_dtypes == [torch.float32, torch.float32]
     assert torch.equal(factor_values[1], factor_values[0])
     for legacy, compact in zip(outputs[0], outputs[1]):
         assert torch.allclose(compact, legacy, atol=8e-3, rtol=1e-2)
